requirements.txt
langchain
google-generativeai
streamlit
python-dotenv
langchain_community
